Created a clear trigger on click
Created a clear trigger to clear forms fields when clicking on account
and clicking out of account.

Also created a new nav called DRAFT so when players click on draft rank
or draft list the rank nav will disappear and the DRAFT nav will show.
This is done for ease of scrolling